+++ src/main/java/com/juanbenevento/wms/inventory/application/port/in/command/AllocateStockCommand.java
@@ -1,4 +1,6 @@
 package com.juanbenevento.wms.inventory.application.port.in.command;
 
-public record AllocateStockCommand(String sku, Double quantity) {
+import java.math.BigDecimal;
+
+public record AllocateStockCommand(String sku, BigDecimal quantity) {
 }
