+++ src/main/java/com/juanbenevento/wms/inventory/domain/event/InventoryAdjustedEvent.java
@@ -1,12 +1,13 @@
 package com.juanbenevento.wms.inventory.domain.event;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record InventoryAdjustedEvent(
         String lpn,
         String productSku,
-        Double oldQuantity,
-        Double newQuantity,
+        BigDecimal oldQuantity,
+        BigDecimal newQuantity,
         String reason,
         String locationCode,
         String username,
