+++ src/main/java/com/juanbenevento/wms/warehouse/application/port/in/usecases/SuggestLocationUseCase.java
@@ -1,5 +1,7 @@
 package com.juanbenevento.wms.warehouse.application.port.in.usecases;
 
+import java.math.BigDecimal;
+
 public interface SuggestLocationUseCase {
-    String suggestBestLocation(String sku, Double quantity);
+    String suggestBestLocation(String sku, BigDecimal quantity);
 }
